Enforce singleton Slot implementation.
As I have been developing @wry/context and its Slot implementation, I have
come to realize how confusing/disastrous it would be if more than one
implementation of the Slot class were used simultaneously, thanks (for
example) to conflicting npm package version constraints.

This is a deep and well known problem in the npm ecosystem, where packages
are never guaranteed to be installed only once, but usually are. This
uncertainty makes the instanceof operator unreliable for classes defined
by npm packages, for example. In the particular case of the Slot class,
the risk of duplication means that bind and noContext cannot reliably
govern all Slots, since there may be more than one currentContext in play.

Unfortunately, there is no way in JavaScript for multiple copies of the
same package to identify themselves to one another secretly, since that
would require some sort of unforgeable key, which cannot be agreed upon
without a preexisting communication channel. This implies we must use
public global state to reconcile multiple @wry/context implementations.
Because it is public, this state must not expose any sensitive data, and
should be immutable if possible. It would have been nice, for example, to
store currentContext globally and allow multiple implementations of the
Slot class to interact with a single currentContext, but then malicious
code could tamper with currentContext, which is unacceptable.

Instead, after much consideration, I have decided to store a single
implementation of the Slot class itself, which makes sense because anyone
can import the Slot class from any copy of the @wry/context package, so
Slot by definition is not sensitive. As a happy side-consequence, this
means `instanceof Slot` should always be reliable.

Because the first Slot class to be defined globally wins, I don't see any
way to do any semantic versioning of the Slot singleton. Unfortunately,
this means we must regard the Slot implementation as essentially frozen
(in all observable ways), since it would be dangerous to rely on any
features added in the future. The additional code in the context.ts module
can change as freely as we like, but the slot.ts core is off-limits.

On the bright side, I believe the Slot class can be safely *subclassed*,
as long as the subclass does not need direct access to currentContext.
